requirements.txt
# ──── Core ML ────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.11.0

# ──── Data Visualization ─────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# ──── Web Application ────────────────────────────────────
streamlit>=1.28.0
altair>=5.0.0

# ──── MLOps ──────────────────────────────────────────────
mlflow>=2.9.0
pyyaml>=6.0.0

# ──── Supporting ─────────────────────────────────────────
python-dateutil>=2.8.0
pillow>=10.0.0
requests>=2.31.0
protobuf>=4.25.0
